Moved inputdtos to subpackage
