+++ src/main/kotlin/no/nb/bikube/core/model/inputDto/TitleInputDto.kt
@@ -1,4 +1,4 @@
-package no.nb.bikube.core.model
+package no.nb.bikube.core.model.inputDto
 
 import java.time.LocalDate
