+++ src/test/kotlin/no/nb/bikube/core/service/CreationValidationServiceTest.kt
@@ -1,7 +1,7 @@
 package no.nb.bikube.core.service
 
 import no.nb.bikube.core.exception.BadRequestBodyException
-import no.nb.bikube.core.model.ItemInputDto
+import no.nb.bikube.core.model.inputDto.ItemInputDto
 import org.junit.jupiter.api.Assertions
 import org.junit.jupiter.api.Test
 import org.junit.jupiter.api.assertThrows
